refactor: add "ui" folder& update build process
