+++ src/components/ui/checkbox.ts
@@ -1,5 +1,5 @@
 import { createEle } from "./createEle";
-import { postMessageToUI } from "../utils/message";
+import { postMessageToUI } from "../../utils/message";
 
 type CheckboxProps = {
 	parentElement: string;
